service implementation almost done
